# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Shradha Oil Center ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reward-schemes:
    get:
      operationId: listRewardSchemes
      tags: [rewards]
      summary: List reward schemes
      responses:
        "200":
          description: Schemes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RewardScheme"

    post:
      operationId: createRewardScheme
      tags: [rewards]
      summary: Create reward scheme
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RewardSchemeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RewardScheme"
  /reward-schemes/{id}:
    patch:
      operationId: updateRewardScheme
      tags: [rewards]
      summary: Update scheme
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RewardSchemeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RewardScheme"

    delete:
      operationId: deleteRewardScheme
      tags: [rewards]
      summary: Delete scheme
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /reward-progress:
    get:
      operationId: listRewardProgress
      tags: [rewards]
      summary: List reward progress for customers
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Progress list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RewardProgress"
  /reward-progress/{id}/disburse:
    post:
      operationId: disburseReward
      tags: [rewards]
      summary: Settle and disburse reward (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Disbursed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RewardProgress"

  # MANUFACTURING
...
components:
  schemas:
...
    RewardScheme:
      type: object
      required: [id, productId, targetLiters, rewardType, rewardValue, startDate, endDate, isActive]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: ["string", "null"]
        targetLiters:
          type: number
        rewardType:
          type: string
          enum: [free_gift, cash_discount, percentage_cashback]
        rewardValue:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
    RewardSchemeInput:
      type: object
      required: [productId, targetLiters, rewardType, rewardValue, startDate, endDate]
      properties:
        productId:
          type: integer
        targetLiters:
          type: number
        rewardType:
          type: string
          enum: [free_gift, cash_discount, percentage_cashback]
        rewardValue:
          type: string
        startDate:
          type: string
        endDate:
          type: string
    RewardSchemeUpdate:
      type: object
      properties:
        targetLiters:
          type: number
        rewardType:
          type: string
        rewardValue:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        isActive:
          type: boolean
    RewardProgress:
      type: object
      required: [id, schemeId, customerId, litersAchieved, isRewardAchieved, isDisbursed]
      properties:
        id:
          type: integer
        schemeId:
          type: integer
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        productName:
          type: ["string", "null"]
        targetLiters:
          type: number
        litersAchieved:
          type: number
        progressPct:
          type: number
        isRewardAchieved:
          type: boolean
        isDisbursed:
          type: boolean
        rewardType:
          type: string
        rewardValue:
          type: string
        schemeEndDate:
          type: string
...