# lib/api-zod/src/generated/types/rewardSchemeUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Shradha Oil Center ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface RewardSchemeUpdate {
  targetLiters?: number;
  rewardType?: string;
  rewardValue?: string;
  startDate?: string;
  endDate?: string;
  isActive?: boolean;
}